+++ EssentialFeed/EssentialFeed/Feed API/HTTPClient.swift
@@ -1,0 +1,17 @@
+//
+//  HTTPClient.swift
+//  EssentialFeed
+//
+//  Created by Sebastian Leon on 29.12.2021.
+//
+
+import Foundation
+
+public enum HTTPClientResult {
+	case success(Data, HTTPURLResponse)
+	case failure(Error)
+}
+
+public protocol HTTPClient {
+	func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
+}
