+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedLoader.swift
@@ -7,15 +7,6 @@
 
 import Foundation
 
-public enum HTTPClientResult {
-	case success(Data, HTTPURLResponse)
-	case failure(Error)
-}
-
-public protocol HTTPClient {
-	func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
-}
-
 public final class RemoteFeedLoader {
 	
 	private let url: URL
